Split up Account component into small components
